Use very pretty fixed-width integer aliases instead of very ugly `_t` types.
Also remove `cstddef` includes unless necessary (ie. for `NULL`; `size_t` isn't special, might as well use `u64`)

+++ kernel/src/bitmap.h
@@ -1,16 +1,15 @@
 #ifndef LENSOR_OS_BITMAP_H
 #define LENSOR_OS_BITMAP_H
 
-#include <cstddef>
-#include <stdint.h>
+#include "integers.h"
 
 class Bitmap {
 public:
-	size_t Size;
-	uint8_t* Buffer;
-	bool Get(uint64_t index);
-	bool Set(uint64_t index, bool value);
-	bool operator[](uint64_t index);
+	u64 Size;
+	u8* Buffer;
+	bool Get(u64 index);
+	bool Set(u64 index, bool value);
+	bool operator[](u64 index);
 };
 
 #endif
